+++ spec/support/fake_smoketest.rb
@@ -10,4 +10,16 @@ module Kapnismology
       self.class.result
     end
   end
+
+  class FakeSmokeTest2 < SmokeTest
+    class << self
+      attr_accessor :name
+      attr_accessor :result
+      attr_accessor :tags
+    end
+
+    def result
+      self.class.result
+    end
+  end
 end